Refactored sentry soruce files in order to make it a bit easier to work with
