+++ src/sentry_advanced.c
