+++ src/sentry_advanced.h
