+++ src/sentry_connect.h
